requirements/local.txt
-r ./base.txt

django-extensions==2.2.5
ipython==7.11.1
